backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "faker>=40.5.1",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "mcp[cli,ws]>=1.27.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]
